Refactor, introduce sampler lambdas
